Implementado recurso para alterar Funcionário.

+++ ponto-backend/src/main/java/org/ponto/DTO/FuncionarioDTO.java
@@ -2,7 +2,7 @@ package org.ponto.DTO;
 
 public class FuncionarioDTO {
 
-    private Integer id;
+    private Long id;
     private String nome;
     private Integer usuarioId;
     // Construtor vazio (pode ser omitido se não for necessário)
@@ -10,17 +10,17 @@ public class FuncionarioDTO {
     }
 
     // Construtor com parâmetros
-    public FuncionarioDTO(Integer id, String nome, Integer usuarioId) {
+    public FuncionarioDTO(Long id, String nome, Integer usuarioId) {
         this.id = id;
         this.nome = nome;
         this.usuarioId = usuarioId;
     }
 
     // Métodos getter e setter para id
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
